included flint2's taylor shift for basering = ZZ[x] and Frac(ZZ[x])
this is a bit unclean, branching on the sage version. also, it connects
to the cython code via "load", which may fail to find the required file
if ore_algebra.py is loaded from a location other than ".".

but the speedup is quite nice, a factor of 100 already for polynomials
of moderate size.