simplify the bounds on rational sequences
Replace the bounds on sequences defined by almost everywhere rational functions
(SeqBounds & co) by a significantly simpler variant. The new bounds are more
readable, easier to implement (especially in the presence of interval
coefficients) and optimize, and less costly to compute. Experimentally, they
appear to be reasonably tight to very tight when the indicial equation
only has small roots (in particular, at ordianry points). However, unlike the
old version which should be tight even in pathological cases, they can also
be quite loose. Thus, the old implementation may still be useful, if only for
comparison purposes--but it would be too much of a hurdle to maintain it for
now, I'd rather add it back in the future if necessary.