bounds: explicit shift (exponent) in HyperexpMajorant
It is important to handle the x^(large) factor separately when computing series
expansions and bounds, because otherwise the Taylor shift done by
HyperexpMajorant.series() densifies numerators of large valuation. And since we
represent the numerators by dense polynomials in order to use arb, we might as
well store the shift separately.

(It might make sense in the future to keep a separate shift starting with in the
computation of residuals, and revert to dense polynomials for the dense part of
the residual and derived quantities.)