binsplit: fix(?) bugs when ordrec < orddeq
For now, we artificially increase the order, because our current
strategy of using vectors of coeffs of the form [u(n-s'), ..., u(n+r-1)]
to make the initial values as simple as possible would break otherwise.

But note that when s' < 0, the functions 1, x, ..., x^(-s'+1) are local
solutions. In the longer term, it would probably be better to always
normalize the recurrence to Sn-valuation zero, and add the columns
corresponding to 1, x, ... in partial_sums().