analytic: start using a custom subclass of diff op internally
The idea is that it will hopefully make it easier to enforce constraints
on operators to which the analytic continuation code applies in a
consistent way, to avoid recomputing things like singularities, and more
generally to keep track of some “global” information. Only a small part
of the code has been ported to the new model at this point.