StoppingCriterion: fail fast by default
Fixes an issue introduced with d84723b4: failing to set fail_fast to
true would cause the code to try improving terrible bounds for
veeerrrryyy long before the intervals blow up sufficiently badly for it
to give up. With the new code structure where we update the fail_fast
flag at each iteration depending on the working precision, it seems
reasonable to set the flag to true by default.