naive_sum: fix experimental interval squashing code
Due to the presence of polynomial parts, the lower bounds on majorants
used in the running error analysis used to have zero terms, making the
round-off error bound infinite and causing the sum to be automatically
recomputed without interval squashing. We now use a separate lower
bound for the terms that could be zero.

If I remember correctly, in earlier versions of the implementation, this
issue was masked by the fact that max(nan, x) was x.