binsplit: fix incorrect propagation of error bounds
An error term intended for the value of the subseries of a log-series
and its derivatives was incorrectly applied to the numerator only. Doing
so effectively divides the contribution of that error term to the radius
of the output by the denominator (which in this context is typically a
large integer).

This is a serious bug that could potentially lead to incorrect results.

However, it is not clear how to construct an actual example where that
would happen. Indeed, the computation of series truncation points is
unaffected and typically ensures that the contribution of the tail is
smaller than the radius of the interval representing the partial sum.
By the time the incorrect error term accounting for the tail is added to
that interval, the interval is likely to contain the correct result
already.

This bug was found while experimenting with p-adic evaluation points. In
the p-adic case, integer denominators are very small, thus the bug leads
to large overestimations instead of incorrect bounds.