monodromy: use formal monodromy in connection method
Use the formal monodromy matrix instead of non-standard branches of the
logarithm to compute monodromy matrices with a base point by the
singular connection method.

This should be about 2× faster when the indicial polynomial has small
dispersion. However, the current version computes the formal monodromy
independently from the connection matrix, and may hence be slower than
the old code for indicial polynomials of extremely large dispersion
(large enough that the computation of the formal monodromy should use
binary splitting).