monodromy: fix sharing of formal monodromy among algebraic conjugates
This fixes an issue in the way the algebraic part of the computation of
the formal monodromy matrix at an algebraic singularity is reused for
its Galois conjugates (9fdc4c).

_formal_monodromy_from_critical_monomials() is able to apply the correct
embedding for each singularity to the critical monomials, but not to the
local exponents; and unfortunately, it is not obvious with the current
code structure how to keep track of the algebraic construction of the
exponents so as to be able to follow them along the embedding. As a
stopgap solution, we restrict the reuse of previously computed formal
monodromy matrices to the case where all exponents are rational.

Thanks to Alexandre Goyer for the bug report and test case.