SA: split and rework contribution_single_singularity()
Split contribution_single_singularity() into smaller functions; simplify
some of the logic by using existing facilities from oa. Also makes it a
bit faster and fixes a bug that led to an incorrect coefficient in one
of the examples. (The bug is *probably* that the computation of local
expansions was performed to insufficient precision in one place.)