SA: fix computation of coefficients of log^k, k ≥ 2
(bug likely introduced while trying to simplify the code in the last few
commits)

Also update the Canham example so that it tests both for that and for
the correct handling of terms with a nonnegative power of (1-z). Note
that in this example, the local exponents are such that log^2 factors
start appearing just at the time when the (1-z)^n becomes nonsingular,
so that the degree in log(n) remains 1.