analytic: custom class for polynomial roots
Introduce a custom class for representing complex roots of polynomials
over number fields; use it for local exponents.

The new class is mainly intended for cases where one manipulates all the
roots of a given irreducible polynomial.

It stores several representations of the same algebraic number (computed
in a lazy way). The previous version of the code used embedded number
field elements (easily convertible to both QQbar and numeric fields)
instead. However, conversions from general embedded number fields to
numeric fields are quite slow, mainly due to the need to convert the
image in QQbar of the generator. In addition, relative number fields do
not support embeddings, forcing us to represent algebraic numbers as
elements of _absolute_ number fields.

In the case of local exponents, the main benefit of the change is that
we can now delay delay the decision to extend the constant field of the
differential operator to represent the exponents, and only do it if we
decided to use an exact recurrence operator.