path: avoid useless exact computation of step increment
The main benefit is to avoid introducing unnecessary extensions when the
endpoints are elements of two distinct number fields.

Plus some reorganization and cleanup.

One doctest output changes in a surprising way. I don't understand
exactly how the inexact input was interpreted before. It now gives only
~53 bits of accuracy in the output. I am replacing the example by a
similar one with an exact input to be able to test the computation with
eps=2^(-100)...