analytic: new handling of quadratic number fields
Sage internally represents elements of all quadratic number fields using
√discriminant as a primitive element. This means that, in general, the
apparent denominator of an element may not be the denominator of the
internal representation.

In order to avoid hidden denominators, we used to normalize the
representation of elements of quadratic fields *of discriminant 1 mod 4*
by changing the primitive element so that the internal and external
representation coincide. However, this was not sufficient, and in some
circumstances (only in calls to local_basis_expansions() and similar
“symbolic” functions?), we could still end up manipulating elements with
hidden denominators.

This commits completely gets rid of the good_number_field() function
used to normalize algebraic extensions. Instead, steps where we need to
“clear denominators” in some expression now use the denominator of the
internal representation in the special case of quadratic fields.

The new approach should work in all cases, allows the extensions present
in user input to be preserved more often, and seems to be slightly
faster.