SA: fix+improve error terms in "exact" cases
When all the expansions we compute are exact, do not force an error term
of the form cst·n^(β-1-order)·log^k(n). Instead, put a suitable
exponentially small O() or B() term.

In passing, fix incorrect construction of the exponentially small term
in cases where this was already supported.