fix serious bug in the handling of branch cuts
Asking for the connection matrix between a non-real singular point s and
a point close to the branch cut of log(s+x) could return a result
corresponding to the wrong branch of the canonical solution at s. In
particular, connection matrices between singular points with the same
imaginary part (other that 0) were sometimes incorrect.

We fix this by being more careful in the choice of intermediate points.
Support for evaluation of singular solutions at balls crossed by a
branch cut is now explicitly limited to the case where no subdivision is
necessary. (It would be possible in principle to support the general
case, but this would complicate the structure of the analytic
continuation code.) The code for computing bounds on P-recursive
sequences used to depend on the ability to evaluate on balls overlapping
branch cuts, but only in a inessential way, and removing this dependency
turns out to lead to better bounds in some cases.