stokes: connection matrices as in the paper (incomplete)
Align the algorithm for computing the connection matrices with the
version in the paper. The new version uses a definition adapted to
Stokes matrices everywhere, whereas, previously, the code used the
standard convention of oaa for computing the connection matrices and
converted them to the other convention afterwards.

As of this commit, though, the “base case” matrices (those corresponding
to edges of the spanning tree and their inverses) are not computed
correctly.