fix handling of singular steps ending on branch cut
In some cases, the nonsingular endpoint of an analytic continuation step
starting at a singular point and ending on the local branch cut would be
replaced by an interval approximation overlapping the cut, leading to
evaluations covering two branches of log(x). This in turn means that the
computed connection matrix would not converge to a point matrix as
prec→∞, and possibly even not become invertible for large prec.