try to avoid computing the singularities exactly...
...in ODE integration problems when a numerical enclosure is enough.

This version is only meant to work in “easy” cases, typically when the
operator has coefficients in ℚ[x] and the singular points can well
separated from each other and from the path using naive interval
arithmetic at 53-bit precision. Even in this case, I am not 100% sure if
it reliably always avoids the exactification or only delays it.